requirements.txt

torchvision==0.2.2
tensorflow==2.4.0
keras==2.4.3
aitextgen==0.3.0
cdqa==1.3.9
streamlit==0.74.1
elasticsearch==7.10.1
tqdm==4.56.0
# transformers==3.5.0
pytorch-pretrained-bert==0.6.2
